requirements.txt
appdirs==1.4.2
astroid==1.4.9
Django==1.10.6
djangorestframework==3.5.4
drf-extensions==0.3.1
drfdocs==0.0.11
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
packaging==16.8
psycopg2==2.7.1
pylint==1.6.5
PyMySQL==0.7.10
pyparsing==2.1.10
six==1.10.0
wrapt==1.10.8
